requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.11
django-decouple==2.1
django-heroku==0.3.1
django-phonenumber-field==7.3.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.7
multidict==6.0.5
packaging==24.0
phonenumbers==8.13.34
phonenumberslite==8.13.34
psycopg2==2.9.9
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
twilio==9.0.4
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
